refactor(buck): separate router and service layers
Extract business logic from router to service layer to achieve clear
separation of concerns. This refactoring improves code maintainability,
testability, and follows the existing project architecture patterns.

- Create BuckService in jupiter/src/service/buck_service.rs
- Add MonoApiService methods as bridge layer
- Add comprehensive configuration validation
- Implement type-safe error handling with BuckError enum
- Add transaction support with ConnectionTrait
- Use OwnedSemaphorePermit for async-safe concurrency control
- Extract batch query concurrency calculation to use Storage-provided method
- Optimize Buck session cleanup task with graceful shutdown support

+++ common/src/config.rs
@@ -687,7 +687,7 @@ pub struct BuckConfig {
 
     /// Maximum concurrent uploads (default: 5) - returned to client as suggestion
     #[serde(default = "default_max_concurrent_uploads")]
-    pub max_concurrent_uploads: u8,
+    pub max_concurrent_uploads: u32,
 
     /// Global upload concurrency limit (default: 50)
     /// Controls the total number of concurrent upload requests
@@ -727,7 +727,7 @@ fn default_max_file_size() -> String {
 fn default_max_files() -> u32 {
     1000
 }
-fn default_max_concurrent_uploads() -> u8 {
+fn default_max_concurrent_uploads() -> u32 {
     5
 }
 fn default_upload_concurrency_limit() -> u32 {
